[GEOT-6759] Remove un-necessary casts from code, add a QA rule to enforce it

+++ modules/extension/graph/src/main/java/org/geotools/graph/util/geom/GeometryUtil.java
@@ -218,8 +218,8 @@ public class GeometryUtil {
             y += c.y;
         }
 
-        x /= (double) n;
-        y /= (double) n;
+        x /= n;
+        y /= n;
 
         LineString simple =
                 gf().createLineString(
