+++ modules/extension/xsd/xsd-wcs/src/main/java/org/geotools/gml4wcs/bindings/GridTypeBinding.java
@@ -105,8 +105,8 @@ public class GridTypeBinding extends AbstractComplexBinding {
         grid.setDimension(BigInteger.valueOf(2));
         grid.setLimits(
                 new GridEnvelope2D(
-                        (int) limitsEnvelope.getLow(0), (int) limitsEnvelope.getLow(1),
-                        (int) limitsEnvelope.getHigh(0), (int) limitsEnvelope.getHigh(1)));
+                        limitsEnvelope.getLow(0), limitsEnvelope.getLow(1),
+                        limitsEnvelope.getHigh(0), limitsEnvelope.getHigh(1)));
 
         List<Node> axisNames = node.getChildren("axisName");
         if (axisNames != null && !axisNames.isEmpty()) {
